added server jwt auth file
