+++ server/utils/auth.js
@@ -1,39 +1,34 @@
-const jwt = require('jsonwebtoken');
+const jwt = require('jasonwebtoken')
 
-// set token secret and expiration date
-const secret = 'mysecretsshhhhh';
-const expiration = '2h';
+const secret = 'mysecret';
+const expiration = '3h';
 
 module.exports = {
-  // function for our authenticated routes
-  authMiddleware: function (req, res, next) {
-    // allows token to be sent via  req.query or headers
-    let token = req.query.token || req.headers.authorization;
+    authMiddleware: function ({ req }) {
+        let token = req.body.token || req.query.token || req.headers.authorization;
 
-    // ["Bearer", "<tokenvalue>"]
-    if (req.headers.authorization) {
-      token = token.split(' ').pop().trim();
-    }
+        // returns the token part after the split and pop
+        if (req.headers.authorization) {
+            token = token.split(' ').pop().trim();
+          }
 
-    if (!token) {
-      return res.status(400).json({ message: 'You have no token!' });
-    }
+        if (!token) {
+        return req;
+        }
 
-    // verify token and get user data out of it
-    try {
-      const { data } = jwt.verify(token, secret, { maxAge: expiration });
-      req.user = data;
-    } catch {
-      console.log('Invalid token');
-      return res.status(400).json({ message: 'invalid token!' });
-    }
+        // tries to verify token and add the decoded user's data to the request
+        try {
+            const { data } = jwt.verify(token, secret, { maxAge: expiration });
+            req.user = data;
+        } catch {
+            console.log('Invalid token');
+        }
 
-    // send to next endpoint
-    next();
-  },
-  signToken: function ({ username, email, _id }) {
-    const payload = { username, email, _id };
-
-    return jwt.sign({ data: payload }, secret, { expiresIn: expiration });
-  },
-};
+        // return the request object so it can be passed to the resolver as `context`
+        return req;
+    },
+    signToken: function ({ email, name, _id }) {
+        const payload = { email, name, _id };
+        return jwt.sign({ data: payload }, secret, { expiresIn: expiration });
+    },
+}
